backend/requirements.txt
# ── Django Core ─────────────────────────────────────────────
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-axes==6.1.1
django-filter==23.5
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-prometheus==2.3.1
django-storages==1.14.2
django-silk==5.0.4

# ── Database ─────────────────────────────────────────────────
psycopg2-binary==2.9.9
pgvector==0.2.4

# ── Cache / Queue ────────────────────────────────────────────
redis==5.0.1
celery==5.3.6
flower==2.0.1

# ── AI / ML ──────────────────────────────────────────────────
langchain==0.1.6
langchain-openai==0.0.6
langchain-community==0.0.19
openai==1.10.0
tiktoken==0.5.2
spacy==3.7.2
transformers==4.36.2
sentence-transformers==2.3.1
torch==2.1.2
keybert==0.8.3
yake==0.4.8
langdetect==1.0.9
datasketch==1.6.4

# ── Web Scraping ─────────────────────────────────────────────
Scrapy==2.11.0
beautifulsoup4==4.12.3
playwright==1.40.0
httpx==0.25.2
requests==2.31.0
feedparser==6.0.11
newspaper3k==0.2.8
trafilatura==1.6.4
html2text==2020.1.16
fake-useragent==1.4.0

# ── Document Generation ───────────────────────────────────────
reportlab==4.0.9
python-pptx==0.6.23
python-docx==0.8.11
fpdf2==2.7.8
Pillow==10.1.0
Jinja2==3.1.3
matplotlib==3.8.2

# ── Cloud ────────────────────────────────────────────────────
boto3==1.34.14
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ── Email ────────────────────────────────────────────────────
sendgrid==6.11.0

# ── Monitoring ───────────────────────────────────────────────
sentry-sdk==1.39.2
prometheus-client==0.19.0
structlog==23.3.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.0
Markdown==3.5.1
pydantic==2.5.3
gunicorn==21.2.0
